pyproject.toml
[project]
name = "tellama"
description = "A bridge that allows Ollama to join Telegram chats as a bot"
authors = [{ name = "K4YT3X", email = "i@k4yt3x.com" }]
dependencies = ["python-telegram-bot>=21.10", "ollama>=0.4.7", "loguru>=0.7.3"]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/k4yt3x/tellama"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "file"
path = "tellama/__init__.py"
